apps/ai-backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Models and Libraries
openai==1.3.7
anthropic==0.7.7
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0
bitsandbytes==0.41.3
langchain==0.0.352
langchain-community==0.0.7
langchain-openai==0.0.2
ollama==0.1.7

# Open Source Models
sentence-transformers==2.2.2
huggingface-hub==0.19.4
tokenizers==0.15.0

# Database and Caching
asyncpg==0.29.0
redis==5.0.1
aioredis==2.0.1
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# HTTP and API clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0
python-json-logger==2.0.7

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Task Queue and Background Jobs
celery==5.3.4
redis==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Model serving and optimization
vllm==0.2.5
llama-cpp-python==0.2.23
ctransformers==0.2.27

# Monitoring and metrics
psutil==5.9.6
GPUtil==1.4.0
py3nvml==0.2.7